+++ src/dummylib/Drawpool.cxx
@@ -1,16 +1,18 @@
-#include "Drawpool.h"
-
-namespace Dummylib{
-
-template<typename T>
-inline
-Drawpool<T>::Drawpool(std::vector<T> iData){
-    this->mPool = iData;
-    srand (time(NULL));
-}
-
-T Drawpool::drawElement(){
-    return mPool.first();
-}
-
-}//namespace Dummylib
+// #include "Drawpool.h"
+//
+// namespace Dummylib{
+//
+// template<typename T>
+// inline
+// Drawpool<T>::Drawpool(std::vector<T> iData){
+//     this->mPool = iData;
+//     srand (time(NULL));
+// }
+//
+// template<typename T>
+// inline
+// T Drawpool<T>::drawElement(){
+//     return mPool.first();
+// }
+//
+// }//namespace Dummylib
