+++ src/dummylib/Drawpool.h
@@ -1,14 +1,21 @@
 #pragma once
 
 #include <vector>
+#include <stdlib.h>
+#include <time.h>
 
 namespace Dummylib{
-    template<class T>
-    class Drawpool{
-    public:
-        Drawpool(std::vector<T>);
-        T drawElement();
-    private:
-        std::vector<T> mPool;
-    };
+template<class T>
+class Drawpool{
+private:
+    std::vector<T> mPool;
+public:
+    Drawpool(std::vector<T> iData){
+        this->mPool = iData;
+        srand (time(NULL));
+    }
+    T drawElement(){
+        return mPool[0];
+    }
+};
 }
